feat(ancillary): implement booking ancillary page with order details and service selection
